+++ public/admin.php
@@ -548,8 +548,8 @@ function clearCache() {
 	console.log("purging cookies/cache");
     cookies = ['filter', 'mtime', 'maxdepth', 'hide_thresh', 'path', 'use_count', 'show_files', 'sort', 'sortorder',
                 'sort2', 'sortorder2', 'resultsize', 'index', 'index2', 'running_task_id', 'tagsshowuntagged', 
-                'tagsshowfiles', 'tagsshowdirectories', 'tagsshowall', 'showotherfiles', 'qumulo', 'minhardlinks', 
-                'mindupes', 'PHPSESSID'];
+                'tagsshowfiles', 'tagsshowdirectories', 'tagsshowall', 'showotherfiles', 'qumulo', 's3', 'minhardlinks', 
+                'mindupes', 'min_change_percent', 'show_new_dirs', 'PHPSESSID'];
     for (var i = 0; i < cookies.length; i++) {
         deleteCookie(cookies[i]);
     }