+++ public/simple.php
@@ -252,6 +252,7 @@ if (isset($_REQUEST['submitted'])) {
     </script>
 <div id="loading">
   <img id="loading-image" width="32" height="32" src="images/ajax-loader.gif" alt="Updating..." />
+  <div id="loading-text"></div>
 </div>
 <iframe name="hiddeniframe" width=0 height=0 style="display:none;"></iframe>
 </body>